creates lexemes folder and moves related artifacts
