+++ libraries/Pliant/Grammars/ILexerRule.cs
@@ -1,0 +1,10 @@
+﻿using Pliant.Tokens;
+
+namespace Pliant.Grammars
+{
+    public interface ILexerRule : ISymbol
+    {
+        IGrammar Grammar { get; }
+        TokenType TokenType { get; }
+    }
+}
