--- libraries/Pliant/ILexeme.cs
@@ -1,17 +1,0 @@
-﻿using Pliant.Tokens;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Pliant
-{
-    public interface ILexeme
-    {
-        string Capture { get; }
-        bool Scan(char c);
-        bool IsAccepted();
-        TokenType TokenType { get; }
-    }
-}
