--- libraries/Pliant/ILexerRule.cs
@@ -1,11 +1,0 @@
-﻿using Pliant.Grammars;
-using Pliant.Tokens;
-
-namespace Pliant
-{
-    public interface ILexerRule : ISymbol
-    {
-        IGrammar Grammar { get; }
-        TokenType TokenType { get; }
-    }
-}
