+++ libraries/Pliant/Lexemes/ILexeme.cs
@@ -1,0 +1,12 @@
+﻿using Pliant.Tokens;
+
+namespace Pliant.Lexemes
+{
+    public interface ILexeme
+    {
+        string Capture { get; }
+        bool Scan(char c);
+        bool IsAccepted();
+        TokenType TokenType { get; }
+    }
+}
