+++ libraries/Pliant/ParseInterface.cs
@@ -1,4 +1,5 @@
 ﻿using Pliant.Collections;
+using Pliant.Lexemes;
 using Pliant.Tokens;
 using System;
 using System.Collections.Generic;
