+++ libraries/Pliant/ParseRunner.cs
@@ -1,5 +1,6 @@
 ﻿using Pliant.Collections;
 using Pliant.Grammars;
+using Pliant.Lexemes;
 using Pliant.Tokens;
 using System.Collections.Generic;
 using System.IO;
