+++ tests/Pliant.Tests.Unit/LexemeTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Pliant.Grammars;
+using Pliant.Lexemes;
 using System.Collections.Generic;
 using System.Linq;
