+++ tests/Pliant.Tests.Unit/TerminalLexemeTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Pliant.Grammars;
+using Pliant.Lexemes;
 using Pliant.Tokens;
 
 namespace Pliant.Tests.Unit
